human_in_the_loop/pyproject.toml
[project]
name = "human-in-the-loop"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiosqlite>=0.22.1",
    "google-adk>=1.25.1",
    "greenlet>=3.3.2",
    "pandas>=3.0.1",
    "rich>=14.3.3",
    "tabulate>=0.9.0",
]

